EMBR-11208 refactor(sessions): move SpanSession* paths to UserSession*/SessionPart*
