--- packages/web-sdk/src/api-sessions/manager/NoOpSpanSessionManager/index.ts
@@ -1,1 +1,0 @@
-export { NoOpSpanSessionManager } from './NoOpSpanSessionManager.ts';
