+++ packages/web-sdk/src/api-sessions/manager/NoOpUserSessionManager/index.ts
@@ -1,0 +1,1 @@
+export { NoOpSpanSessionManager } from './NoOpUserSessionManager.ts';
