--- packages/web-sdk/src/api-sessions/manager/ProxySpanSessionManager/index.ts
@@ -1,1 +1,0 @@
-export { ProxySpanSessionManager } from './ProxySpanSessionManager.ts';
