+++ packages/web-sdk/src/api-sessions/manager/ProxyUserSessionManager/index.ts
@@ -1,0 +1,1 @@
+export { ProxySpanSessionManager } from './ProxyUserSessionManager.ts';
