+++ packages/web-sdk/src/managers/EmbraceUserSessionManager/README.md
