+++ packages/web-sdk/src/managers/EmbraceUserSessionManager/constants.ts
