+++ packages/web-sdk/src/managers/EmbraceUserSessionManager/index.ts
@@ -1,4 +1,4 @@
-export { EmbraceSpanSessionManager } from './EmbraceSpanSessionManager.ts';
+export { EmbraceSpanSessionManager } from './EmbraceUserSessionManager.ts';
 export type {
   SpanSessionManagerInternal,
   UserSessionConfig,
