+++ packages/web-sdk/src/managers/EmbraceUserSessionManager/types.ts
