+++ packages/web-sdk/src/managers/EmbraceUserSessionManager/utils/activity.ts
