+++ packages/web-sdk/src/managers/EmbraceUserSessionManager/utils/index.ts
