+++ packages/web-sdk/src/managers/EmbraceUserSessionManager/utils/state.ts
