+++ packages/web-sdk/src/managers/index.ts
@@ -10,13 +10,13 @@ export { EmbracePageManager } from './EmbracePageManager/index.ts';
 export type { EmbracePageManagerArgs } from './EmbracePageManager/types.ts';
 export type { SDKFeaturesManager } from './EmbraceSDKFeaturesManager/index.ts';
 export { EmbraceSDKFeaturesManager } from './EmbraceSDKFeaturesManager/index.ts';
-export type {
-  SpanSessionManagerInternal,
-  UserSessionConfig,
-} from './EmbraceSpanSessionManager/index.ts';
-export { EmbraceSpanSessionManager } from './EmbraceSpanSessionManager/index.ts';
 export {
   EmbraceExtendedSpan,
   EmbraceTraceManager,
 } from './EmbraceTraceManager/index.ts';
 export { EmbraceUserManager } from './EmbraceUserManager/index.ts';
+export type {
+  SpanSessionManagerInternal,
+  UserSessionConfig,
+} from './EmbraceUserSessionManager/index.ts';
+export { EmbraceSpanSessionManager } from './EmbraceUserSessionManager/index.ts';
