+++ packages/web-sdk/src/processors/IdentifiableSessionLogRecordProcessor/types.ts
@@ -1,4 +1,4 @@
-import type { SpanSessionManagerInternal } from '../../managers/EmbraceSpanSessionManager/index.ts';
+import type { SpanSessionManagerInternal } from '../../managers/EmbraceUserSessionManager/index.ts';
 
 export interface IdentifiableSessionLogRecordProcessorArgs {
   spanSessionManager: SpanSessionManagerInternal;
